New feature: #87 support Android 12+
