+++ pkg/util/ebpf/bpf.go
@@ -1,54 +1,10 @@
 package ebpf
 
 import (
-	"bufio"
 	"fmt"
 	"golang.org/x/sys/unix"
 	"os"
 )
-
-const (
-	BOOT_CONFIG_PATH   = "/boot/config-%s"
-	CONFIG_BTF_TAGNAME = "CONFIG_DEBUG_INFO_BTF"
-)
-
-func IsEnableBTF() (bool, error) {
-	found, e := checkKernelBTF()
-	if e == nil && found {
-		return true, nil
-	}
-
-	i, e := getOSUnamer()
-	if e != nil {
-		return false, e
-	}
-	bootConf := fmt.Sprintf(BOOT_CONFIG_PATH, i.Release)
-
-	// Open file bootConf.
-	f, err := os.Open(bootConf)
-	if err != nil {
-		return false, err
-	}
-	defer f.Close()
-
-	var KernelConfig = make(map[string]string)
-	s := bufio.NewScanner(f)
-	if err := parse(s, KernelConfig); err != nil {
-		return false, err
-	}
-	bc, found := KernelConfig[CONFIG_BTF_TAGNAME]
-	if !found {
-		// 没有这个配置项
-		return false, nil
-	}
-
-	//如果有，在判断配置项的值
-	if bc != "y" {
-		// 没有开启
-		return false, nil
-	}
-	return true, nil
-}
 
 type UnameInfo struct {
 	SysName    string
@@ -88,13 +44,13 @@ func charsToString(ca [65]byte) string {
 	return string(s[0:lens])
 }
 
-// from internal/btf/btf.go
+// from internal/btf/bpf.go
 // checkKernelBTF attempts to load the raw vmlinux BTF blob at
 // /sys/kernel/btf/vmlinux and falls back to scanning the file system
 // for vmlinux ELFs.
 
 func checkKernelBTF() (bool, error) {
-	_, err := os.Stat("/sys/kernel/btf/vmlinux")
+	_, err := os.Stat(SYS_KERNEL_BTF_VMLINUX)
 
 	// if exist ,return true
 	if err == nil {
@@ -111,17 +67,6 @@ func findVMLinux() (bool, error) {
 		return false, err
 	}
 	release := kv.Release
-	// use same list of locations as libbpf
-	// https://github.com/libbpf/libbpf/blob/9a3a42608dbe3731256a5682a125ac1e23bced8f/src/btf.c#L3114-L3122
-	locations := []string{
-		"/boot/vmlinux-%s",
-		"/lib/modules/%s/vmlinux-%[1]s",
-		"/lib/modules/%s/build/vmlinux",
-		"/usr/lib/modules/%s/kernel/vmlinux",
-		"/usr/lib/debug/boot/vmlinux-%s",
-		"/usr/lib/debug/boot/vmlinux-%s.debug",
-		"/usr/lib/debug/lib/modules/%s/vmlinux",
-	}
 
 	for _, loc := range locations {
 		_, err := os.Stat(fmt.Sprintf(loc, release))
